add TomographyFitter classes
